requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic[email]==2.9.0
geopy==2.4.1
better-profanity==0.7.0
httpx[http2]==0.27.0
PyJWT==2.9.0
slowapi==0.1.9
alembic==1.13.2
apscheduler==3.10.4
boto3==1.35.0
pytest==8.3.3
pytest-asyncio==0.24.0
